fix: BoardService와 BoardViewService 분리
